refactor api urls: make more restful
