+++ english_diary/diaries/api/diary/__init__.py
@@ -1,3 +1,2 @@
 from .detail import DiaryDetailAPIView
 from .create import DiaryCreateAPIView
-from .delete import DiaryDeleteAPIView
