+++ english_diary/english_diary/templates/home.html
@@ -42,7 +42,7 @@ $( document ).ready(function() {
         if($("form")[0].checkValidity()) {
             var diaryContent = $('#diary-content').val();
             var diaryDatetime = $('#diary-datetime').val();
-            var diaryCreateAPIUrl = "/api/diary/create/";
+            var diaryCreateAPIUrl = "/api/diary/";
             var data = {
                 datetime: diaryDatetime,
                 content: diaryContent
@@ -71,7 +71,7 @@ $( document ).ready(function() {
     // Delete Diary
     $('#diary-delete').click(function() {
         var diaryDatetime = $('#diary-datetime').val();
-        var diaryDeleteAPIUrl = "/api/diary/" + diaryDatetime + "/delete/";
+        var diaryDeleteAPIUrl = "/api/diary/" + diaryDatetime;
         var data = {
             datetime: diaryDatetime,
         };
